remove general request mapping, add openapi description, add service name list method
