+++ InquirerClient/src/main/java/caller/ClientApplication.java
@@ -1,0 +1,10 @@
+package caller;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ClientApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(ClientApplication.class, args);
+	}
+}
